Auxiliary Functions

A set of additional functions that handle various subsidiary tasks, such as the conversion of various datatype in different floating point arithmetics or the generation of synthetic problems.

Conversion of datatypes

XGMRES.xconvert — Function

a::u = xconvert(
    u::DataType,  # Target floating point arith for the conversion
    num::T        # Real number to convert
) where {T<:AbstractFloat}

a::Vector{u} = xconvert(
    u::DataType,    # Target floating point arith for the conversion
    vec::Vector{T}  # Vector of real to convert
) where {T<:AbstractFloat}

a::Matrix{u} = xconvert(
    u::DataType,    # Target floating point arith for the conversion
    mat::Matrix{T}  # Matrix of real to convert
) where {T<:AbstractFloat}

a::IncompleteLU.ILUFactorization{u,Int64}) = xconvert(
    u::DataType,                                 # Target floating point arith for the conversion
    mat::IncompleteLU.ILUFactorization{T,Int64}  # IncompleteLU factors structure to convert
) where {T<:AbstractFloat}

a::LU{u,Matrix{u},Vector{Int64}}) = xconvert(
    u::DataType,                         # Target floating point arith for the conversion
    fact::LU{T,Matrix{T},Vector{Int64}}  # LinearAlgebra.LU factors structure to convert
) where {T<:AbstractFloat}

a::Symmetric{u,SparseMatrixCSC{u,Int64}} = xconvert(
    u::DataType,                                # Target floating point arith for the conversion
    mat::Symmetric{T,SparseMatrixCSC{T,Int64}}  # Symmetric sparse matrix to convert
) where {T<:AbstractFloat}

a::SparseMatrixCSC{u,Ti} = xconvert(
    u::DataType,                 # Target floating point arith for the conversion
    mat::SparseMatrixCSC{Tv,Ti}  # Sparse matrix to convert
) where {Tv<:AbstractFloat,Ti<:Integer}

a::Adjoint{u,SparseMatrixCSC{u,Int64}} = xconvert(
    u::DataType,                              # Target floating point arith for the conversion
    mat::Adjoint{T,SparseMatrixCSC{T,Int64}}  # Adjoint sparse matrix to convert
) where {T<:AbstractFloat}

a::Vector{Vector{u}} = xconvert(
    u::DataType,            # Target floating point arith for the conversion
    vec::Vector{Vector{T}}  # Vector of vector of real to convert
) where {T<:AbstractFloat}

a::AlgebraicMultigrid.MultiLevel{
        AlgebraicMultigrid.Pinv{u},
        GaussSeidel{SymmetricSweep},
        GaussSeidel{SymmetricSweep},
        SparseMatrixCSC{u,Int64}, 
        Adjoint{u,SparseMatrixCSC{u,Int64}},
        SparseMatrixCSC{u,Int64}, 
        AlgebraicMultigrid.MultiLevelWorkspace{Vector{u},1}
} = xconvert(
    u::DataType,  # Target floating point arith for the conversion
    mat::AlgebraicMultigrid.MultiLevel{         # AlgebraicMultigrid datatype
                AlgebraicMultigrid.Pinv{T},     # to convert
                GaussSeidel{SymmetricSweep},
                GaussSeidel{SymmetricSweep},
                SparseMatrixCSC{T,Int64}, 
                Adjoint{T,SparseMatrixCSC{T,Int64}},
                SparseMatrixCSC{T,Int64}, 
                AlgebraicMultigrid.MultiLevelWorkspace{Vector{T},1}}
) where {T<:AbstractFloat}

a::ILUZero.ILU0Precon{u,Int64,u}) = xconvert(
    u::DataType,                        # Target floating point arith for the conversion
    mat::ILUZero.ILU0Precon{T,Int64,T}  # IncompleteLU factors structure to convert
) where {T<:AbstractFloat}

𝑃::Poly = xconvert(
    u::DataType,       # Target floating point arith for the conversion
    poly::Poly{TA,TP}  # Polynomial to convert
) where {TA<:AbstractFloat,TP<:AbstractFloat}

Random dense problems generator

XGMRES.gen_mat — Function

A::Matrix{Float64} = gen_mat(
    n       ::Integer,              # Dimension of the matrix to generate 
    κₐ      ::Float64,              # Condition number of the matrix to generate
    opt     ::String   = "default"  # If set to "reproductible", the same Randome.seed!() is set and the same sequence of matrices is generated accross Julia sessions
)

Generate a random dense matrix with given condition number.

XGMRES.gen_mat_with_prec — Function

A::Matrix{Float64}, M::Matrix{Float64} = gen_mat_with_prec(
    n    ::Integer,           # Dimension of the matrix A and M to generate
    κₐ   ::Float64,           # Condition number of the matrix A
    κₘ   ::Float64,           # Condition number of the preconditioner M
    opt  ::String="default"   # If set to "reproductible", the same Randome.seed!() is set and the same sequence of matrices is generated accross Julia sessions
)

Generate a random dense matrix A and an associated preconditioner M that satisfy

\[\kappa(M^{-1} A) = \kappa(A M^{-1}) = \kappa(A) / \kappa(M).\]