XGMRES

The xgmres solver, combined with a given preconditioner, can make use of up to six different floating point arithmetics. For the sake of readability, we use the following set of consistent notations to identify easily the precision at which a given variable is stored and an operation is performed. These notations are consistent accross this documentation and the source code. You can find below a table listing the different operations, their associated floating point datatype, and the subscript used for this floating point datatype.

Operations	Floating point datatype	Subscript

Residual $Ax_{i+1} - b$	$u_r$	r

Update $x_{i+1} = x_{i} + d_{i}$	$u$	no subscript

Preconditioner computation	$u_x$	x

Preconditioner application $M\times v$	$u_m$	m

Matrix application $A\times v$	$u_a$	a

Rest of the GMRES operations	$u_s$	s

For instance, if a variable is noted varₓ, this variable is stored in precision $u_x$ and is likely associated to the preconditioner computation.

The floating point arithmetic $u$ is also refered to as the working precision. This is the arithmetic at which the solution x of xgmres is delivered. We also use this notation when it is not pertinent to make a distinction between different floating point arithmetics. Similarly, we do not add subscript to a variable in those cases.

Finally, in some parts of the code we refer to variables explicitly stored in Float64 or Float128 (through the Quadmath package). For those cases we use the following subscripts.

Floating point datatype	Subscript

Float64	e

Floar128	t

Note that we would have liked subscripts d and q, respectively for Float64 (double precision) and Float128 (quadruple precision), but Julia's unicode does not have these options.

XGMRES.xgmres — Function

x::Vector{u}, stats::Dict = xgmres(
    A         ::AbstractMatrix{TA},                  
    b         ::AbstractVector{TB};                  
    precond   ::Preconditioner            = nothing, 
    kind      ::String                    = "left",  
    xexact    ::Union{Vector{TX},Nothing} = nothing, 
    maxrestrt ::Union{Integer,   Nothing} = nothing, 
    m         ::Union{Integer,   Nothing} = nothing, 
    τ         ::Union{Float64,   Nothing} = nothing, 
    bstop     ::Union{Float64,   Nothing} = nothing, 
    fstop     ::Union{Float64,   Nothing} = nothing, 
    stop      ::Union{Float64,   Nothing} = nothing, 
    verbose   ::Bool                      = true,    
    do_stats  ::Bool                      = false,   
    do_κ      ::Bool                      = false,   
    u         ::DataType                  = Float64, 
    uᵣ        ::DataType                  = Float64, 
    uₛ        ::DataType                  = Float64, 
    maxmem    ::Float64                   = 3*1e9    
) where {TA<:AbstractFloat,TB<:AbstractFloat,TX<:AbstractFloat}

Solve the linear system $Ax=b$ with mixed precision restarted preconditioned GMRES, where $A\in\mathbb{R}^{n\times n}$ and $x,b \in\mathbb{R}^n$. The function requires three precision parameters and repeats the three following steps until convergence:

Compute the residual $r_i = b-Ax_i$ in precision $u_r$

Solve $A d_i = r_i/||r_i||_{\infty}$ with preconditioned GMRES run in precision $u_s$

Update the solution $x_{i+1} = x_i + ||r_i||_{\infty}\times d_i$ in precision $u$

The precisions $u_r$ and $u$ concern, respectively, the precisions at which the residual (step 1) and the update of the solution (step 3) are computed. The precision $u_s$ is the precision at which all the operations of the inner GMRES, except the preconditioned matrix-vector product, are computed (step 2). The preconditioned matrix-vector product should be provided through the precond parameter.

Preconditioned matrix-vector product

The preconditioned matrix-vector product can use its own set of precisions and does not have to be implemented in precision $uₛ$. One should use the Preconditioner structure provided by the package to implement it or use provided implementations by the package.

The algorithm is composed of two imbricated loops: the outer restart loop and the inner GMRES iterations. The restart loop is stopped if:

We reached the maximum number of restart iterations maxrestrt.

We observe no significant error reduction over three consecutives iterations. This is quantified by the parameter stop which defines the maximum acceptable ratio between the forward errors of two consecutive iterations.

We reached the expected accuracies on the forward and backward errors defined by the parameters fstop and bstop.

If the parameter xexact is not provided, the forward error is not computed, and the above stopping criterion are solely based on the computed backward error. The backward and forward error of the linear system are computed, respectively, as follows:

\[\frac{||Ax_i-b||_{\infty}}{||A||_{\infty}||x_i||_{\infty} + ||b||_{\infty}},\]

\[\frac{||x_i - x||_{\infty}}{||x||_{\infty}}.\]

The inner GMRES loop is stopped if:

We reached the maximum number of GMRES iterations m.

The backward error of the preconditioned system $Ad_i=r_i$ falls below a given tolerance defined by the paramter τ.

The backward error of the preconditioned system is computed as follows:

\[\frac{||M^{-1}(Ad_{i,j} - r_i/||r_i||_{\infty})||_{\infty}}{||M^{-1}r_i/ ||r_i||_{\infty}||_{\infty}} \qquad \text{(left-preconditioned)}\]

\[\frac{||Ad_{i,j} - r_i/||r_i||_{\infty})||_{\infty}}{||r_i/ ||r_i||_{\infty}||_{\infty}} \qquad \text{(right-preconditioned)}\]

Right-preconditioned backward error

The backward error of the right-preconditioned system is the backward error of the original system $A d_i = r_i/||r_i||_{\infty}$.

Through the parameter do_stats, it can be asked to the function to compute the backward errors of the original system $Ax=b$ for each iteration of GMRES. The forward errors are also computed if the exact solution of the system is provided through the parameter xexact. The condition numbers of the preconditioner and the preconditioned matrix can also be computed through another parameter do_κ. It lets the possibility to compute the do_stats advanced errors on systems where the condition number is too expensive or impossible to compute.

Stats mode

The stat mode increases substantially the resource comsumption and is only relevant for numerical study of the algorithm. It is not intended to be used in practice.

Input arguments

A: A square general matrix $A\in\mathbb{R}^{n\times n}$

b: A right-hand side $b\in\mathbb{R}^n$

Keyword arguments

precond: A Preconditioner object embedding the linear operators used for preconditioning

kind: Set the kind of preconditioning ("left", "right", or "flexible")

xexact: Exact solution of the system. If provided the forward error is computed in stat mode.

maxrestrt: Maximum number of restart iterations.

m: Maxmum number of inner GMRES iterations.

τ: Stopping criterion for the inner iterations of GMRES based on the backward error of the correction system $A d_i = r_i/||r_i||_{\infty}$.

bstop: Stopping criterion for the outer restart iterations based on the backward error of the original system $A x_i = b$.

fstop: Stopping criterion for the outer restart iterations based on the forward error of the original system.

stop: Stopping criterion for the outer restart iterations based on the based on the non-improvement of the forward error over 3 iterations. If the forward error is not improved by a ratio higher than stop for three consecutive restart iterations, the algorithm is stopped.

verbose: Information and convergence logs are displayed at each iteration.

do_stats: Activate the stats mode and get many more logs.

do_κ: Compute the condition numbers of the preconditioner $M$ and the preconditioned matrix $AM$ or $MA$ (depending on the kind of preconditioning used).

u: Precision at which the update $x_{i+1} = x_{i} + d_{i}$ is computed.

uᵣ: Precision at which the residual $Ax_{i+1} - b$ is computed.

uₛ: Precision at which the operations in the inner loop of GMRES are computed.

maxmem: Use to set m if m is not provided in argument. It chooses m such that the memory used to store the Krylov bases is lower than maxmem.